requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
Pillow==10.3.0
